new file structure + justfile collection command
