+++ backend/ETL/dockerfile
@@ -16,10 +16,13 @@ RUN uv run python -c "import duckdb; duckdb.connect().execute('INSTALL spatial')
 
 # The source data collection scripts (source --> destination)
 COPY --chown=app:app ./data_collection /code/data_collection
+# The master script file to run data collection
+COPY --chown=app:app ./run_data_collection.py /code/run_data_collection.py
+
 
 ENV DATA_DIR=/data
 ENV PYTHONPATH=/code
 
 # Command to run the data collection scripts (container entrypoint) 
-# TODO: Add a year argument to the collectioncommand
-ENTRYPOINT ["uv", "run", "python", "-m", "data_collection.run"]
+# TODO: Add a year argument to the collection command (not yet)
+ENTRYPOINT ["uv", "run", "python", "-m", "run_data_collection"]
