+++ backend/data_collection_OLD/economic.py
@@ -20,7 +20,7 @@ B23001 prime-age variable codes (structure: 7 vars per age-sex group):
 Output: vt_acs5_b_economic_tidy.parquet
 """
 
-from ETL.data_collection.base import YEARS, VarGroup, run_acs_b_scrape
+from data_collection_OLD.base import YEARS, VarGroup, run_scrape
 
 SL = "Labor Force"
 SI = "Income"
@@ -70,11 +70,10 @@ fetch_specs = {
     "B19301": ["B19301_001E"],
 }
 
-
-def collect():
+if __name__ == "__main__":
     import argparse
 
-    from ETL.data_collection.base import ALL_GEOS
+    from data_collection_OLD.base import ALL_GEOS
 
     p = argparse.ArgumentParser(description="Scrape ACS B-table economic data.")
     p.add_argument(
@@ -92,7 +91,7 @@ def collect():
     )
     args = p.parse_args()
     selected_geos = [(k, *ALL_GEOS[k]) for k in args.geos]
-    df = run_acs_b_scrape(
+    run_scrape(
         fetch_specs,
         var_groups,
         "vt_acs5_b_economic_tidy.parquet",
@@ -100,9 +99,3 @@ def collect():
         selected_geos,
         append=args.append,
     )
-
-    return df
-
-
-if __name__ == "__main__":
-    df = collect()
