+++ backend/tests/test_base.py
@@ -6,7 +6,7 @@ Tests for data_collection/base.py:
 import pandas as pd
 import pytest
 
-from data_collection.base import VarGroup, compute_tidy_generic, pct
+from data_collection_OLD.base import VarGroup, compute_tidy_generic, pct
 
 # ---------------------------------------------------------------------------
 # pct()
